Bring the QHY_AGENT_DEV branch up to MASTER organization
Plus various cleaning

	modified:   AbstractAgents/CompositeAgent.py
	modified:   AbstractAgents/SpecialAgent.py
	modified:   DTO/DTO.py
	deleted:    HardwareClients.py
	modified:   IndiClient.py
	modified:   Locutus/Locutus.py
	renamed:    QHY600Camera.py -> QHY600-INDI-Composite/QHY600Camera.py
	renamed:    QHY600CcdCooler.py -> QHY600-INDI-Composite/QHY600CcdCooler.py
	renamed:    QHY600Composite.py -> QHY600-INDI-Composite/QHY600Composite.py
	renamed:    qhy600_agent.yaml -> QHY600-INDI-Composite/QHY600_composite_config.yaml
	new file:   QHY600-INDI-Composite/qhy600_agent.yaml
	modified:   SBIG-INDI-Composite/SBIGCamera.py
	modified:   SBIG-INDI-Composite/SBIGCcdCooler.py
	modified:   SBIG-INDI-Composite/SBIGComposite.py
	modified:   SBIG-INDI-Composite/SBIGFilterWheel.py
	new file:   lorax/utils.py
