+++ QHY600-INDI-Composite/QHY600_composite_config.yaml
@@ -1,8 +1,4 @@
 ---
-log_file: qhy600_composite.log
-broker_hosts:
-  - tanagra
-  - 61613
 agents_in_composite:
   - camera1:
       agent_name: QHY600Camera
@@ -12,6 +8,12 @@ agents_in_composite:
       incoming_topic: lorax.ldtboresight.dto.camera1
       outgoing_topic: lorax.ldtboresight.camera1.broadcast
       dto_command_topic: lorax.ldtboresight.camera1.dto
+      status:
+        - CCD_FRAME_TYPE
+        - CCD_BINNING
+        - CCD_GAIN
+        - CCD_EXPOSURE
+        - CCD_DIRECTORY_LOCATION
   - ccdcooler1:
       agent_name: QHY600CcdCooler
       cooler_name: "QHY CCD QHY600M-d0a5a44"
@@ -20,4 +22,11 @@ agents_in_composite:
       incoming_topic: lorax.ldtboresight.dto.ccdcooler1
       outgoing_topic: lorax.ldtboresight.ccdcooler1.broadcast
       dto_command_topic: lorax.ldtboresight.ccdcooler1.dto
+      status:
+        - CCD_COOLER
+        - CCD_TEMPERATURE
+broker_hosts:
+  - tanagra
+  - 61613
+log_file: qhy600_composite.log
 message_wait_time: 0.5
